+++ examples/readme.md
@@ -1,0 +1,17 @@
+# Examples
+
+Minimal example of each function of csgrs.
+
+Here, we do not use any shared data, so we'll bind the generic S to ().
+
+# Scenes
+
+Additional “SCENES” Demonstrating Each Function Minimally, these start with a capital letter
+
+In these scenes, we typically:
+1) Create the shape
+2) Extrude (if 2D) so we can save an STL
+3) Optionally union with a small arrow that points to a location of interest in the shape
+4) Save the result as an STL, e.g. "scene_XX_something.stl"
+
+Because many shapes are already in the main examples, these scenes are just short examples to help with explanation.
