+++ src/main.rs
@@ -158,149 +158,6 @@ fn main() {
     
     let crescent2d = CSG::crescent(10.0, 7.0, 4.0, 64, None);
     let _ = fs::write("stl/crescent2d.stl", crescent2d.to_stl_ascii(""));
-
-    // ---------------------------------------------------------
-    // Additional “SCENES” Demonstrating Each Function Minimally
-    //
-    // In these scenes, we typically:
-    //   1) Create the shape
-    //   2) Extrude (if 2D) so we can save an STL
-    //   3) Optionally union with a small arrow that points to
-    //      a location of interest in the shape
-    //   4) Save the result as an STL, e.g. "scene_XX_something.stl"
-    //
-    // Because many shapes are already shown above, these are
-    // just short examples to help with explanation.
-    // ---------------------------------------------------------
-
-    // Scene A: Demonstrate a right_triangle(width=2, height=1)
-    {
-        let tri_2d = CSG::right_triangle(2.0, 1.0, None);
-        // A tiny arrow pointing from the right-angle corner outward:
-        let arrow = CSG::arrow(
-            Point3::new(0.0, 0.0, 0.1), // at corner
-            Vector3::new(0.5, 0.0, 0.0), 
-            8,
-            true,
-            None::<()>,
-        )
-        .scale(0.05, 0.05, 0.05);
-        let scene = tri_2d.extrude(0.1).union(&arrow);
-        let _ = fs::write("stl/scene_right_triangle.stl", scene.to_stl_ascii("scene_right_triangle"));
-    }
-
-    // Scene B: Demonstrate extrude_vector(direction)
-    {
-        let circle2d = CSG::circle(1.0, 32, None);
-        // extrude along an arbitrary vector
-        let extruded_along_vec = circle2d.extrude_vector(Vector3::new(0.0, 0.0, 2.0));
-        let _ = fs::write("stl/scene_extrude_vector.stl", extruded_along_vec.to_stl_ascii("scene_extrude_vector"));
-    }
-
-    // Scene E: Demonstrate center() (moves shape so bounding box is centered on the origin)
-    {
-        let off_center_circle = CSG::circle(1.0, 32, None).translate(5.0, 2.0, 0.0).extrude(0.1);
-        let centered_circle = off_center_circle.center();
-        let _ = fs::write("stl/scene_circle_off_center.stl", off_center_circle.to_stl_ascii("scene_circle_off_center"));
-        let _ = fs::write("stl/scene_circle_centered.stl", centered_circle.to_stl_ascii("scene_circle_centered"));
-    }
-
-    // Scene F: Demonstrate float() (moves shape so bottom is at z=0)
-    {
-        let sphere_for_float = CSG::sphere(1.0, 16, 8, None).translate(0.0, 0.0, -1.5);
-        let floated = sphere_for_float.float();
-        let _ = fs::write("stl/scene_sphere_before_float.stl", sphere_for_float.to_stl_ascii("scene_sphere_before_float"));
-        let _ = fs::write("stl/scene_sphere_floated.stl", floated.to_stl_ascii("scene_sphere_floated"));
-    }
-
-    // Scene G: Demonstrate inverse() (flips inside/outside)
-    {
-        // Hard to visualize in STL, but let's do it anyway
-        let inv_sphere = sphere.inverse();
-        #[cfg(feature = "stl-io")]
-        let _ = fs::write("stl/scene_inverse_sphere.stl", inv_sphere.to_stl_binary("scene_inverse_sphere").unwrap());
-    }
-
-    // Scene H: Demonstrate tessellate() (forces triangulation)
-    {
-        let tri_sphere = sphere.tessellate();
-        #[cfg(feature = "stl-io")]
-        let _ = fs::write("stl/scene_tessellate_sphere.stl", tri_sphere.to_stl_binary("scene_tessellate_sphere").unwrap());
-    }
-
-    // Scene I: Demonstrate slice(plane) – slice a cube at z=0
-    {
-        let plane_z = Plane::from_normal(Vector3::z(), 0.5 );
-        let sliced_polygons = cube.slice(plane_z);
-        let _ = fs::write("stl/scene_sliced_cube.stl", cube.to_stl_ascii("sliced_cube"));
-        // Save cross-section as well
-        let _ = fs::write("stl/scene_sliced_cube_section.stl", sliced_polygons.to_stl_ascii("sliced_cube_section"));
-    }
-
-    // Scene J: Demonstrate re-computing vertices() or printing them
-    {
-        let circle_extruded = CSG::circle(1.0, 32, None).extrude(0.5);
-        let verts = circle_extruded.vertices();
-        println!("Scene J circle_extruded has {} vertices", verts.len());
-        // We'll still save an STL so there's a visual
-        let _ = fs::write("stl/scene_j_circle_extruded.stl", circle_extruded.to_stl_ascii("scene_j_circle_extruded"));
-    }
-
-    // Scene K: Demonstrate reuleaux_polygon with a typical triangle shape
-    // (already used sides=4 above, so let's do sides=3 here)
-    {
-        let reuleaux_tri = CSG::reuleaux(3, 2.0, 16, None).extrude(0.1);
-        let _ = fs::write("stl/scene_reuleaux_triangle.stl", reuleaux_tri.to_stl_ascii("scene_reuleaux_triangle"));
-    }
-
-    // Scene L: Demonstrate rotate_extrude (360 deg) on a square
-    {
-        let small_square = CSG::square(1.0, 1.0, None).translate(2.0, 0.0, 0.0);
-        let revolve = small_square.rotate_extrude(360.0, 24);
-        let _ = fs::write("stl/scene_square_revolve_360.stl", revolve.to_stl_ascii("scene_square_revolve_360"));
-    }
-
-    // Scene M: Demonstrate “mirror” across a Y=0 plane
-    {
-        let plane_y = Plane::from_normal(Vector3::y(), 0.0);
-        let shape = CSG::square(2.0, 1.0, None).translate(1.0, 1.0, 0.0).extrude(0.1);
-        let mirrored = shape.mirror(plane_y);
-        let _ = fs::write("stl/scene_square_mirrored_y.stl", mirrored.to_stl_ascii("scene_square_mirrored_y"));
-    }
-
-    // Scene N: Demonstrate scale() 
-    {
-        let scaled = sphere.scale(1.0, 2.0, 0.5);
-        #[cfg(feature = "stl-io")]
-        let _ = fs::write("stl/scene_scaled_sphere.stl", scaled.to_stl_binary("scene_scaled_sphere").unwrap());
-    }
-
-    // Scene O: Demonstrate transform() with an arbitrary affine matrix
-    {
-        use nalgebra::{Matrix4, Translation3};
-        let xlate = Translation3::new(2.0, 0.0, 1.0).to_homogeneous();
-        // Scale matrix
-        let scale_mat = Matrix4::new_scaling(0.5);
-        // Combine
-        let transform_mat = xlate * scale_mat;
-        let shape = CSG::cube(1.0,1.0,1.0, None).transform(&transform_mat);
-        let _ = fs::write("stl/scene_transform_cube.stl", shape.to_stl_ascii("scene_transform_cube"));
-    }
-
-    // Scene P: Demonstrate offset(distance)
-    {
-        let poly_2d = CSG::polygon(
-            &[
-                [0.0,0.0],
-                [2.0,0.0],
-                [1.0,1.5]
-            ],
-            None
-        );
-        let grown = poly_2d.offset(0.2);
-        let scene = grown.extrude(0.1);
-        let _ = fs::write("stl/scene_offset_grown.stl", scene.to_stl_ascii("scene_offset_grown"));
-    }
     
     let gear_involute_2d = CSG::involute_gear_2d(
         2.0,   // module [mm]
